finishing exception

+++ src/main/java/com/aws/demo/service/GerenciadorTarefasService.java
@@ -7,6 +7,9 @@ import org.springframework.stereotype.Service;
 
 import com.aws.demo.Status.TarefasStatusEnum;
 import com.aws.demo.entity.Tarefa;
+import com.aws.demo.excecoes.NaoPermitidoAlterarStatusException;
+import com.aws.demo.excecoes.NaoPermitirExcluirException;
+import com.aws.demo.excecoes.TarefaExistenteException;
 import com.aws.demo.repository.GerenciadorTarefasRepository;
 
 import jakarta.transaction.Transactional;
@@ -23,7 +26,14 @@ public class GerenciadorTarefasService {
 	@Autowired
 	private GerenciadorTarefasRepository gereciadorTarefasRepository;
 	
+	//criar
 	public Tarefa salvarTarefa(CadastrarTarefaRequest request) {
+		
+		Tarefa tarefaValidacao = gereciadorTarefasRepository.findByTItuloOrDescricao(request.getTitulo(), request.getDescricao());
+		
+		if (tarefaValidacao != null) {
+			throw new TarefaExistenteException("Ja existe uma tarefa com o mesmo titulo ou descricao");
+		}
 		
 		Tarefa tarefa = new Tarefa.TarefaBuilder()
 				.setQuantidadeHorasEstimadas(request.getQuantidadeHorasEstimadas())
@@ -47,18 +57,39 @@ public class GerenciadorTarefasService {
 	
 	public Tarefa atualizarTarefa(Long id, AtualizarTarefaRequest request){
 		
-		Tarefa tarefa = new Tarefa.TarefaBuilder()
-				.setQuantidadeHorasEstimadas(request.getQuantidadeHorasEstimadas())
-				.setStatus(request.getStatus())
-				.setTitulo(request.getTitulo())
-				.setDescricao(request.getDescricao())
-				.setResponsavel(request.getResponsavel())
-				.build();
+		Tarefa tarefa = this.gereciadorTarefasRepository.findById(id).get();
+		
+		if (tarefa.getStatus().equals(TarefasStatusEnum.FINALIZADA)) {
+			throw new NaoPermitidoAlterarStatusException("Nao permitido mover a tarefa que esta FINALIZADA");
+		}
+		
+		if (tarefa.getStatus().equals(TarefasStatusEnum.CRIADA) && request.getStatus().equals(TarefasStatusEnum.FINALIZADA)) {
+			throw new NaoPermitidoAlterarStatusException("Nao permitido mover a tarefa para FINALIZADA se ela estiver com status de CRIADA");
+		}
+		
+		if (tarefa.getStatus().equals(TarefasStatusEnum.BLOQUEADA) && request.getStatus().equals(TarefasStatusEnum.FINALIZADA)) {
+			throw new NaoPermitidoAlterarStatusException("Nao permitido mover a tarefa para FINALIZADA se ela estiver com status de BLOQUEADA");
+		}
+		
+		tarefa.setQuantidadeHorasEstimadas(request.getQuantidadeHorasEstimadas());
+		tarefa.setStatus(request.getStatus());
+		tarefa.setTitulo(request.getTitulo());
+		tarefa.setDescricao(request.getDescricao());
+		tarefa.setResponsavel(request.getResponsavel());
+		tarefa.setQuantidadeHorasRealizada(request.getQuantidadeHorasRealizada());
 		
 		 return this.gereciadorTarefasRepository.save(tarefa);
 	}
 	
 	public void excluirTarefa(Long id){
+		 
+		 
+		 Tarefa tarefa = this.gereciadorTarefasRepository.findById(id).get();
+		 
+		 if (TarefasStatusEnum.CRIADA.equals(tarefa.getStatus())) {
+			 throw new NaoPermitirExcluirException();
+		 }
+		 
 		 this.gereciadorTarefasRepository.deleteById(id);
 	}
